hotfix: hibernate config

+++ apps/backend/src/main/java/com/intranet/backend/repository/PasswordResetTokenRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Long> {
+public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Integer> {
     Optional<PasswordResetToken> findByToken(String token);
     Optional<PasswordResetToken> findByUserAndUsedFalse(User user);
 }
